getProfile() handler implemented

+++ profile-service/profiles/repositories/profileRepository.go
@@ -11,10 +11,10 @@ type ProfileRepository interface {
 	SaveNew(p models.Profile) (err error)
 	SaveExisting(p models.Profile) (err error)
 
-	DeleteByID(id int64) (err error)
+	DeleteByID(id uint32) (err error)
 
-	FindByID(id int64) (p models.Profile, err error)
+	FindByID(id uint32) (p models.Profile, err error)
 	FindByUsernameAndPassword(username, password string) (p models.Profile, err error)
 
-	GetSeveralOrderByScorePaginated(page, pageSize int32) (p models.Profiles, err error)
+	GetSeveralOrderByScorePaginated(page, pageSize uint32) (p models.Profiles, err error)
 }
